requirements.txt
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
install==1.3.4
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
pandas==1.2.2
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
